requirements.txt
pandas>=1.3.0
numpy>=1.20.0
jupyter>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
pytorch>=1.10.0
torchvision>=0.11.0
mlflow>=1.20.0
fastapi>=0.70.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
minio>=7.0.0
boto3>=1.18.0
pyarrow>=6.0.0 